Improved the AI Settings UI
